ai-backend/requirements.txt
# GENAVATOR1 AI Backend Requirements
# Install with: pip install -r requirements.txt

# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Document Processing (PyMuPDF)
PyMuPDF==1.26.4
python-docx==1.2.0

# Core Python dependencies
pydantic==2.5.0
typing-extensions==4.15.0
lxml==6.0.2
passlib[bcrypt]==1.7.4

# Hugging Face ML libraries
transformers==4.35.2
torch==2.1.1
tokenizers==0.15.0
datasets==2.14.6
sentence-transformers==2.2.2

# Natural language processing
nltk==3.8.1
spacy==3.7.2
textstat==0.7.3

# Document processing libraries (MAIN FOCUS)
PyMuPDF==1.23.8
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
xlsxwriter==3.1.9
pandas==2.1.3

# Text processing and utilities
beautifulsoup4==4.12.2
markdown==3.5.1
python-magic==0.4.27

# Async and HTTP utilities
aiofiles==23.2.1
httpx==0.25.2
requests==2.31.0

# Configuration and environment
pydantic==2.5.0
python-dotenv==1.0.0

# Logging and monitoring
loguru==0.7.2

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: GPU acceleration (uncomment if you have NVIDIA GPU)
# torch-audio==2.1.1
# torchaudio==2.1.1
# accelerate==0.24.1

# Optional: Additional ML models
# sentence-transformers==2.2.2
# sklearn==1.3.2